feat: Add dynamic input rendering for CV output
This commit updates several components to which are all linked to enable the dynamic rendering of user inputs into the CV output container.
This feature allows a user to input data into the form fields and this is immediately rendered in the (yet to be styled) CV output section.

The changes include:

InputBasic.jsx:
- Added controlled components with `value` and `onChange` props, which allows for real-time input handling.

InputSection.jsx:
- Added state management for `formValues` to dynamically capture and update user inputs.
- Added `handleFieldChange` to standardize field handling and dynamically generate state keys from the field labels.

OutputSection.jsx:
- Updated to iterate over `formValues` & render each input field's value, this allows real time rendering based on the users input via the InputSection.

SectionContainer.jsx:
- Adds state management for `formValues` and passes the `handleInputChange` callback down to `InputSection` for input handling.