feat: Add unique ID & improve input event handling
This commit is both a feature and also a fix to the issue of state being duplicated in form fields.

Problem:
- Initially, used array indexes &field labels as keys for form values which led to shared state across different fields with the same label, such as "Description".
This caused an unintended duplication of input values across these fields.

Solution:
- Introducing crypto.randomUUID() to generate unique IDs for each field, which ensures each field's state is managed independently.
This solves the problem of shared state and value duplication across fields with the same label.

Improve Event Handling in Form Fields

Problem:
- Previously I was mistakenly passing the change event handler (onChange) directly from the parent to the child components instead of just extracting the input value.
This caused the entire event object to be stored in state, rather than just the input value causing prop type errors.

Solution:
- Now the code only extract the input value from the event object (e.target.value) in the onChange handler within FormField.