feat: Add profile photo upload capability
Refactored the FileInput component to utilize URL.createObjectURL() for generating a temp URL for uploaded images.

OutputSection Component:
Has been updated to render content based on the type of value received.
If the value is identified as a blob URL (indicating an image), it is rendered within an <img> tag. Text values are displayed as span elements.

Note:
This will be refactored at a later date to create a more visually styled CV but for now this is implemented to verify functionality.