feat: Add 'repeatable' flag for form sections
This commit adds a 'repeatable' flag to identify sections in the input form which can be duplicated by the user pressing the 'add more' button.
This will allow users to add more information to certain sections such as Qualifications, Works Experience and Education.

This feature will be added incrementally in the following commits.