feat: Add ability to add more form sections
In this commit introduces the ability to dynamically add new input sections to allow users to add more content to their CVs.
Previously the user was only able to add one 'work experience' or one 'technical skill'.
Now when the 'add more' button is clicked, a new set of form fields are created within that section.
At the moment its not as refined a UX as I would like, as  but I intend to enhance this feature to be more user friendly.

Changes:
- Extract the logic for adding a new set of fields into a separate file (addFieldSetUtil.js).
- Implement cloneFieldWithNewId function to ensure each new field added has a unique ID.
- Modified InputSection.jsx to use the the new utility function, maintaining state for form values and dynamically added sections more effectively.
- Update InputSectionDisclosure to support the new dynamic addition of fields.
- Added some comments for better readability to make the code more understandable.