refactor: Separate CVHeader into own component
This commit begins building a modular set of components which will create the final CV.

This first change introduces a CVHeader component, which will encapsulate the profile photo, name, and surname elements.
In this initial iteration, only the photo element is included to validate functionality, the name and surname elements will be added in subsequent commits.