feat: Add name & surname to CVHeader component
This commit extends the CVHeader component to include name and surname props.
This change allows the CVHeader component to display the user's name and surname within the CV header, in addition to the user's photo.

The OutputSection component has been updated to extract name and surname from formValues using a new labelsToIds mapping, which links form labels to specific IDs for simpler data handling.
The SectionContainer component now includes a useMemo hook to create a labelsToIds mapping from the formSections structure, which maps labels to UUIDs for more straightforward referencing.

I think this will be a temporary approach however as I don't think it will be very scalable.
I'm currently looking into potentially integrating React Context API for global state management which might offer a better state management solution.

Having said that, it currently functions as expected and the CVHeader component separates out these values to allow them to be worked with and styled easily.