fix: prop type warning for missing onClick
The app was constantly giving a warning regarding a missing prop type.
To fix this I've added a placeholder onClick handler to the Create PDF button within the Header component.
Also added a TODO comment to highlight the intention to implement the actual functionality for PDF creation at a later stage.