feat: Render skills section in OutputSection
This commit adds in the functionality to add the user generated skills inputs and display them in the OutputSection.
This is very much a work in process iteration as I feel the code is quite verbose and needs to be simplified to allow easier integration of the other dynamically generated fields such as work experience.
However for now this has working functionality to serve as a base for future refactoring.
I also plan to add the ability for users to delete the dynamically generated skills but this will be part of a future commit.

Overview of how it works:
- Update `OutputSection` component to handle both initial and dynamically added skill inputs
- Extract initial skill value from `formValues` using `labelsToIds`
- Filter `formValues` entries to get dynamically added skills with keys starting with 'skill-'
- Combine initial skill and dynamic skills into a single `skills` array
- Pass the `skills` array to the `CVSkill` component for rendering
- Modify `CVSkill` component to accept and render an array of skills
- Update `addFieldSetUtil.js` to prepend 'skill-' to the ID of dynamically added skill fields