feat: add CVWorkExperience component
This commit introduces the CVWorkExperience component.
This component takes formValues as a prop, filters and reduces the entries to create workExperienceItems.
Each work experience item is then mapped to render a section with company, position, start and end dates, and a description.
I will revise the styling in future commits, this commit simply introduces the initial functionality.