refactor: CVSkill component to filter skills
This commit updates the CVSkill component to accept the formValues object as a prop and filter out the skills internally.
Previously, the component expected a pre-filtered array of skills as a prop.
This change makes the CVSkill component more self-contained and removes its related logic from the CVOutputSection component which was becoming confusing.