refactor: Simplify labelToIds in SectionContainer
This commit removes the previous useMemo hook in an effort to make the code more readable.
Instead labelsToIds is now defined as a regular object outside the component.
Populating labelsToIds by iterating over the formSections array and mapping field labels to their respective IDs.
Passing labelsToIds directly as a prop to the OutputSection component.