fix: rendering of workexperience input
This is more of a temp fix that needed further investigation. I intended to use 'work-experience' prefix as part of the Ids to make them more meaningful,
however when this is used the work experience data will not render, therefor I've reverted back to the original Ids.
I plan to investigate this further and come up with a better solution when I have more time.