refactor: Use item.id for key in CVWorkExperience
This commit addresses an anti-pattern in the CVWorkExperience component where the index was being used as the key prop when rendering work experience items.
As I understand it, using the index as the key can lead to issues when the order of items changes or when items are added or removed from a list.

Changes made:
- Replaced the use of index with item.id for the key prop in the map function.