feat: Implement data persistence via local storage
The main objective of this commit is to introduce data persistence using local storage.
This allows users to preserve any input data between page refreshes and or navigation away from the app.

Changes:

formStructure.js:
- Removed the dynamic generation of UUIDs for field IDs in the 'Personal Information', 'Summary', 'Technical Skills', 'Qualifications', and 'Education' sections.
- Replaced the dynamically generated UUIDs with fixed field IDs to ensure consistency across page refreshes.

SectionContainer.jsx:
- Added a useEffect hook to load form values from local storage on the initial render.
- If stored form values exist in local storage, they are parsed and used to populate the formValues state using the setFormValues function.
- Added another useEffect hook to save form values to local storage whenever they change.
- The formValues state is stringified and stored in local storage using the localStorage.setItem method.

FormInputSection.jsx:
- Updated the addFieldSet function to generate a new groupId using crypto.randomUUID() when adding a new set of fields to a section.
- The generated groupId is passed to the addNewFieldSet function to ensure unique field IDs for newly added field sets.

Reason for the Changes:
By removing the dynamically generated UUIDs and using fixed field IDs, we ensure that the stored form values in local storage match the field IDs in the form structure.
This allows the form values to be correctly loaded and populated when the page is refreshed or revisited.

The useEffect hooks in the SectionContainer component handle the loading and saving of form values to local storage.
When the component mounts, it checks for stored form values in local storage and populates the formValues state accordingly.
Whenever the formValues state changes, the updated values are saved back to local storage.

The FormInputSection component no longer needs to handle loading form values from local storage, as this responsibility is now delegated to the SectionContainer component.
The addFieldSet function is updated to generate a new groupId for each new set of fields added to a section, ensuring unique field IDs.