feat: Add A4 dimensions to PDF print render
This commit introduces dimensions for the create PDF functionality.
These styles are applied to the CVOutputSection component when the document is printed/saved as a PDF and replicate an A4 size document.

Styles are applied via a a new media query `@media print` in the index.css file.
This class is applied to the CVOutputSection which ensures a width and height equivalent to an A4 paper size (210mm x 297mm) when printed/saved.
Additionally, the `page-break-after: always;` rule is used to force a page break after the CVOutputSection, ensuring that each CVOutputSection starts on a new page when printed.