fix: add label with form field in ImageUploader
The label in the ImageUploader component was not associated with any form field.
This commit fixes this issue by adding an 'id' attribute to the input element and a corresponding 'htmlFor' attribute to the label element, associating the label with the input field.
This fixes a devtools warning about a form control without a label.